+++ ninja/app/src/main/java/com/ninjahome/ninja/utils/ContactIconUtils.kt
@@ -17,7 +17,7 @@ object ContactIconUtils {
     fun getDrawable(fontSize: Int, address: String, subName: String): TextDrawable {
         mDrawableBuilder.fontSize(fontSize)
         val index = ChatLib.iconIndex(address, ColorUtil.colorSize)
-        val iconColor = ColorUtil.colors[index]
+        val iconColor = ColorUtil.colors[index.toInt()]
         val drawable = mDrawableBuilder.textColor(context().getColor(R.color.white)).endConfig().buildRound(subName, context().resources.getColor(iconColor))
         return drawable
     }
